+++ vllm/model_executor/layers/fused_moe/experts/trtllm_nvfp4_moe.py
@@ -17,7 +17,7 @@ from vllm.model_executor.layers.fused_moe.topk_weight_and_reduce import (
     TopKWeightAndReduceNoOP,
 )
 from vllm.model_executor.layers.fused_moe.utils import trtllm_moe_pack_topk_ids_weights
-from vllm.model_executor.layers.quantization.utils.flashinfer_utils import (
+from vllm.model_executor.layers.quantization.utils.flashinfer_moe import (
     activation_to_flashinfer_int,
 )
 from vllm.model_executor.layers.quantization.utils.quant_utils import (
